+++ gtfold/gtfold-mfe/include/data.h
@@ -24,44 +24,44 @@
 
 #include "constants.h"
 
-extern int poppen[5];
-extern int maxpen;
-extern int eparam[11];
-extern int multConst[3]; /* for multiloop penalties. */
-extern int dangle[4][4][4][2]; /* Contain dangling energy values */
-extern int inter[31]; /* Contains size penalty for internal loops */
-extern int bulge[31]; /* Contain the size penalty for bulges */
-extern int hairpin[31]; /* Contains the size penalty for hairpin loops */
-extern int stack[256]; /* Stacking energy used to calculate energy of stack loops */
-extern int tstkh[256]; /* Terminal mismatch energy used in the calculations of hairpin loops */
-extern int tstki[256]; /* Terminal mismatch energy used in the calculations of internal loops */
-extern int tloop[maxtloop + 1][2];
+extern double poppen[5];
+extern double maxpen;
+extern double eparam[11];
+extern double multConst[3]; /* for multiloop penalties. */
+extern double dangle[4][4][4][2]; /* Contain dangling energy values */
+extern double inter[31]; /* Contains size penalty for internal loops */
+extern double bulge[31]; /* Contain the size penalty for bulges */
+extern double hairpin[31]; /* Contains the size penalty for hairpin loops */
+extern double stack[256]; /* Stacking energy used to calculate energy of stack loops */
+extern double tstkh[256]; /* Terminal mismatch energy used in the calculations of hairpin loops */
+extern double tstki[256]; /* Terminal mismatch energy used in the calculations of internal loops */
+extern double tloop[maxtloop + 1][2];
 extern int numoftloops;
-extern int iloop22[5][5][5][5][5][5][5][5]; /* 2*2 internal looops */
-extern int iloop21[5][5][5][5][5][5][5]; /* 2*1 internal loops */
-extern int iloop11[5][5][5][5][5][5]; /* 1*1 internal loops */
-extern int coax[6][6][6][6];
-extern int tstackcoax[6][6][6][6];
-extern int coaxstack[6][6][6][6];
-extern int tstack[6][6][6][6];
-extern int tstkm[6][6][6][6];
-extern int auend;
-extern int gubonus;
-extern int cint; /* cint, cslope, c3 are used for poly C hairpin loops */
-extern int cslope;
-extern int c3;
-extern int efn2a;
-extern int efn2b;
-extern int efn2c;
-extern int triloop[maxtloop + 1][2];
+extern double iloop22[5][5][5][5][5][5][5][5]; /* 2*2 internal looops */
+extern double iloop21[5][5][5][5][5][5][5]; /* 2*1 internal loops */
+extern double iloop11[5][5][5][5][5][5]; /* 1*1 internal loops */
+extern double coax[6][6][6][6];
+extern double tstackcoax[6][6][6][6];
+extern double coaxstack[6][6][6][6];
+extern double tstack[6][6][6][6];
+extern double tstkm[6][6][6][6];
+extern double auend;
+extern double gubonus;
+extern double cint; /* cint, cslope, c3 are used for poly C hairpin loops */
+extern double cslope;
+extern double c3;
+extern double efn2a;
+extern double efn2b;
+extern double efn2c;
+extern double triloop[maxtloop + 1][2];
 extern int numoftriloops;
-extern int init;
+extern double init;
 extern int gail; /* It is either 0 or 1. It is used for grosely asymmetric internal loops */
-extern float prelog;
+extern double prelog;
 
-extern int tstackm[5][5][6][6];
-extern int tstacke[5][5][6][6];
-extern int tstacki23[5][5][5][5];
+extern double tstackm[5][5][6][6];
+extern double tstacke[5][5][6][6];
+extern double tstacki23[5][5][5][5];
 
 
 #define fourBaseIndex(a, b, c, d) (((a) << 6) + ((b) << 4) + ((c) << 2) + (d))
