+++ gtfold/gtfold-mfe/include/global.h
@@ -47,8 +47,8 @@ void print_gtfold_usage_help();
 #ifdef __cplusplus
 int read_sequence_file(const char* filename, std::string& seq);
 bool encodeSequence(string seq);
-void save_ct_file(string outputFile, string seq, int energy) ;
-void save_ct_file(string outputFile, string seq, int energy, int *structure1); 
+void save_ct_file(string outputFile, string seq, double energy) ;
+void save_ct_file(string outputFile, string seq, double energy, int *structure1); 
 #endif
 
 void init_checkPair();
